requirements.txt
beautifulsoup4==4.13.4
openai==1.97.1
python-dotenv==1.1.1
requests==2.32.4
streamlit==1.47.1
